backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==12.0
google-genai==0.3.0
python-multipart==0.0.9
